requirements.txt
Django==5.2.1
django-filter==25.2
djangorestframework==3.16.0
pylint==3.3.7
pylint-django==2.6.1
python-dotenv==1.2.1
requests==2.32.5
